# 백엔드/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HUFS Town API
  version: 0.5.0
  description: GDG HUFS SSO, spaces, invitations, map editing and world admission. Media signaling uses the authenticated world WebSocket contract; Node control routes are private. Never log credentials, invite codes or admission headers.
paths:
...
  /api/v1/spaces/{id}:
    parameters: [{ $ref: "#/components/parameters/SpaceId" }]
    get:
      operationId: getSpace
      security: [{ sessionCookie: [] }]
      description: PRIVATE requires membership; a denied read is indistinguishable from a missing space (404). UNLISTED can be read by ID by any active signed-in account.
      responses:
        "200": { $ref: "#/components/responses/Space" }
        default: { $ref: "#/components/responses/SpaceError" }
    patch:
      operationId: updateSpace
      security: [{ sessionCookie: [], csrfHeader: [] }]
      description: OWNER only. Existing members retain access after visibility changes. Capacity and map template are immutable in this version.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SpaceEdit" }
      responses:
        "200": { $ref: "#/components/responses/Space" }
        default: { $ref: "#/components/responses/SpaceError" }
...
components:
...
  parameters:
    SpaceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: HUFS_TOWN_SESSION }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-TOKEN }
  schemas:
...
    SpaceEdit:
      type: object
      required: [name, description, visibility]
      properties:
        name:
          {
            type: string,
            minLength: 1,
            maxLength: 60,
            description: Trimmed; no control/format characters.,
          }
        description:
          {
            type: string,
            maxLength: 300,
            description: Trimmed single-line description; no control/format characters.,
          }
        visibility: { type: string, enum: [PUBLIC, UNLISTED, PRIVATE] }
        approvalRequired:
          {
            type: boolean,
            default: false,
            description: Only applies to PUBLIC or UNLISTED spaces. Existing members and valid invitees can enter without owner approval.,
          }
...
    Space:
      allOf:
        - { $ref: "#/components/schemas/SpaceEdit" }
        - type: object
          required:
            [
              id,
              capacity,
              templateId,
              role,
              approvalRequired,
              joinRequestStatus,
              favorite,
              archived,
            ]
          properties:
            id: { type: string, format: uuid }
            capacity: { type: integer, minimum: 1, maximum: 100 }
            templateId:
              {
                type: string,
                enum: [CAMPUS_SQUARE, STUDY_SPACE, MEETUP_HALL, OFFICE],
              }
            role: { type: string, enum: [OWNER, ADMIN, MEMBER, ""] }
            approvalRequired: { type: boolean }
            joinRequestStatus:
              { type: string, enum: [PENDING, APPROVED, REJECTED, ""] }
            favorite:
              {
                type: boolean,
                description: Whether the authenticated user saved this space.,
              }
            archived:
              {
                type: boolean,
                description: True when returned by the owner's archived-space view.,
              }
...
    AuthError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
  responses:
...
    Space:
      description: Space metadata and the caller's role.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Space" }
...
    SpaceError:
      description: 400 invalid input/invite/event date; 401 session missing; 403 owner/account/CSRF/approval/access-block denied; 404 missing or private space; 409 quota/request/event/ownership-transfer/archive conflict; 429 per-account rate or daily event limit; 503 storage unavailable. Uses the common code/message shape.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AuthError" }
...